Lots of small improvements to complete chapter 1-- inventory can now be removed, overlays can now invert the image underneath, timers now support sub-second time, and the monster-shooter control now removes whichever opponent has run out of HP; Chapter 1 quest can now be played through to the start of Chapter 2

+++ src/main/java/quest/control/TimerStartControl.java
@@ -18,7 +18,7 @@ public class TimerStartControl extends QuestControl {
     @Override
     public String onExecute(String tag) {
         System.out.println("TimerStartControl: onExecute: tag=" + tag);
-        int seconds = Integer.parseInt(getTagToken(tag, 1, false));
+        double seconds = (double) Double.parseDouble(getTagToken(tag, 1, false));
         String timerName = getTagToken(tag, 2, true);
 
         String eventName = Quest.TIMER_EVENT_PREFIX + ":" + timerName;
